agents/analysis/requirements.txt
# Analysis Agent specific requirements
google-cloud-aiplatform>=1.36.0
google-cloud-bigquery>=3.13.0
google-cloud-firestore>=2.13.0
google-cloud-pubsub>=2.18.0
google-cloud-logging>=3.8.0
pyyaml>=6.0
pandas>=2.0.0
numpy>=1.24.0
psutil>=5.9.0
fastapi>=0.104.0
uvicorn>=0.24.0